Crud for events are working properly and fully ok
