--- frontend/src/components/EditEvent.jsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function EditEvent() {
-  return <div>EditEvent</div>;
-}
