+++ frontend/src/pages/Administration.jsx
@@ -1,13 +1,13 @@
 import React from "react";
 import { useNavigate } from "react-router-dom";
-import CrudEvent from "../components/CrudEvent";
+import AllEvents from "../components/AllEvents";
 
 function Administration() {
   const navigate = useNavigate();
   return (
     <div>
       <p>Page administration</p>
-      <CrudEvent />
+      <AllEvents />
       <button type="button" onClick={() => navigate("/")}>
         Retour à la page d'accueil
       </button>
